requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
SQLAlchemy>=2.0.25,<3.0.0
alembic>=1.13.1,<2.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
